database_config.py
import os
from dotenv import load_dotenv
import psycopg2

load_dotenv()

DB_NAME = os.getenv('DB_NAME')
DB_USER = os.getenv('DB_USER')
DB_PASSWORD = os.getenv('DB_PASSWORD')
DB_HOST = os.getenv('DB_HOST')
DB_PORT = os.getenv('DB_PORT')

def get_curser():
    print(f"Connecting to database at {DB_HOST}:{DB_PORT}")
    connection = psycopg2.connect(
        dbname=DB_NAME,
        user=DB_USER,
        password=DB_PASSWORD,
        host=DB_HOST,
        port=DB_PORT
    )
    # Create a cursor
    cursor = connection.cursor()
    return connection, cursor